Moving jquery helper
