+++ __tests__/__renderer__/classes/CalendarFactory.mjs
@@ -1,6 +1,7 @@
+/* eslint-disable no-undef */
 'use strict';
 
-import '../jquery.mjs';
+import '../../../__mocks__/jquery.mjs';
 
 import assert from 'assert';
 import { stub } from 'sinon';
