+++ __tests__/__renderer__/notification-channel.mjs
@@ -1,6 +1,6 @@
 'use strict';
 
-import './jquery.mjs';
+import '../../__mocks__/jquery.mjs';
 
 import assert from 'assert';
