+++ __tests__/__renderer__/preferences.mjs
@@ -1,7 +1,7 @@
 /* eslint-disable no-undef */
 'use strict';
 
-import './jquery.mjs';
+import '../../__mocks__/jquery.mjs';
 
 import assert from 'assert';
 import fs from 'fs';
